restaurant api created

+++ swiggato/src/main/java/com/example/swiggato/model/Cart.java
@@ -18,7 +18,7 @@ public class Cart {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
-    double cartTotal;
+    int cartTotal;
 
     @OneToOne
     @JoinColumn
